+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Add `strUnique` method
 - Add `strToAscii` method
 - Add `urlJoin` method
+- Add `arrWithDefault` method
 
 ## 0.0.2 (2024-09-10)